+++ changelog.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Transitioned from `aws-sdk` to [`aws-lite`](https://aws-lite.org)
+- Breaking change: `arm64` is now the default Lambda architecture
 - Breaking change: `nodejs20.x` and `python3.12` are now the default Node.js and Python Lambda runtimes, respectively
 - Breaking change: removed support for Node.js 14.x (now EOL, and no longer available to created in AWS Lambda)
 - Added Node.js 20.x to test matrix